+++ backend/src/config/routes.rs
@@ -6,7 +6,7 @@ use actix_web_httpauth::middleware::HttpAuthentication;
 
 use crate::controller::{
     base_layer_image, blossoms, config, guided_tours, layers, map, plant_layer, plantings, plants,
-    seed, sse, users,
+    seed, sse, timeline, users,
 };
 
 use super::auth::middleware::validator;
@@ -38,32 +38,37 @@ pub fn config(cfg: &mut web::ServiceConfig) {
                 .service(map::update)
                 .service(map::update_geometry)
                 .service(
-                    web::scope("/{map_id}/layers")
-                        .service(layers::find)
-                        .service(layers::find_by_id)
-                        .service(layers::create)
-                        .service(layers::delete)
+                    web::scope("/{map_id}")
                         .service(
-                            web::scope("/base/images")
-                                .service(base_layer_image::create)
-                                .service(base_layer_image::update)
-                                .service(base_layer_image::delete),
-                        )
-                        .service(
-                            web::scope("/base/{layer_id}/images").service(base_layer_image::find),
-                        )
-                        .service(
-                            web::scope("/plants")
-                                .service(plant_layer::heatmap)
-                                .service(plant_layer::find_relations)
+                            web::scope("/layers")
+                                .service(layers::find)
+                                .service(layers::find_by_id)
+                                .service(layers::create)
+                                .service(layers::delete)
                                 .service(
-                                    web::scope("/plantings")
-                                        .service(plantings::find)
-                                        .service(plantings::create)
-                                        .service(plantings::update)
-                                        .service(plantings::delete),
+                                    web::scope("/base/images")
+                                        .service(base_layer_image::create)
+                                        .service(base_layer_image::update)
+                                        .service(base_layer_image::delete),
+                                )
+                                .service(
+                                    web::scope("/base/{layer_id}/images")
+                                        .service(base_layer_image::find),
+                                )
+                                .service(
+                                    web::scope("/plants")
+                                        .service(plant_layer::heatmap)
+                                        .service(plant_layer::find_relations)
+                                        .service(
+                                            web::scope("/plantings")
+                                                .service(plantings::find)
+                                                .service(plantings::create)
+                                                .service(plantings::update)
+                                                .service(plantings::delete),
+                                        ),
                                 ),
-                        ),
+                        )
+                        .service(timeline::get_timeline),
                 ),
         )
         .service(
@@ -74,6 +79,7 @@ pub fn config(cfg: &mut web::ServiceConfig) {
         )
         .service(web::scope("/users").service(users::create))
         .service(web::scope("/blossoms").service(blossoms::gain))
+        .service(web::scope("/timeline").service(timeline::get_timeline))
         .wrap(NormalizePath::trim())
         .wrap(auth);