+++ frontend/src/config/i18n/de/index.ts
@@ -12,6 +12,7 @@ import fileSelector from './fileSelector.json';
 import geomap from './geomap.json';
 import guidedTour from './guidedTour.json';
 import imprint from './imprint.json';
+import keybindings from './keybindings.json';
 import landingPage from './landingPage.json';
 import layerSettings from './layerSettings.json';
 import layers from './layers.json';
@@ -19,6 +20,7 @@ import maps from './maps.json';
 import nextcloudIntegration from './nextcloudIntegration.json';
 import overview from './overview.json';
 import plantAndSeedSearch from './plantAndSeedSearch.json';
+import plantRelations from './plantRelations.json';
 import plantSearch from './plantSearch.json';
 import plantingSuggestions from './plantingSuggestions.json';
 import plantings from './plantings.json';
@@ -70,4 +72,6 @@ export default {
   readOnly,
   toolboxTooltips,
   polygon,
+  plantRelations,
+  keybindings,
 };