+++ frontend/src/features/map_planning/layers/plant/actions.ts
@@ -11,16 +11,24 @@ import {
   UpdatePlantingAddDateActionPayload,
   UpdatePlantingRemoveDateActionPayload,
   UpdatePlantingAdditionalNamePayload,
+  UpdatePlantingNoteActionPayload,
 } from '@/api_types/definitions';
 import updateAddDatePlanting, {
   createPlanting,
   deletePlanting,
   movePlanting,
   transformPlanting,
+  updatePlantingNotes,
   updateRemoveDatePlanting,
 } from '../../api/plantingApi';
 import useMapStore from '../../store/MapStore';
 import { Action, TrackedMapState } from '../../store/MapStoreTypes';
+import {
+  decreaseAddedPlantsForDate,
+  increaseAddedPlantsForDate,
+  timlineEventsUpdateAdedDate,
+  timlineEventsUpdateRemoveDate,
+} from '../../utils/TimelineEventsHelper';
 import { filterVisibleObjects } from '../../utils/filterVisibleObjects';
 
 export class CreatePlantAction
@@ -45,6 +53,7 @@ export class CreatePlantAction
 
   apply(state: TrackedMapState): TrackedMapState {
     const timelineDate = useMapStore.getState().untrackedState.timelineDate;
+    increaseAddedPlantsForDate(this._data[0].addDate || timelineDate);
 
     return {
       ...state,
@@ -94,6 +103,15 @@ export class DeletePlantAction
   }
 
   apply(state: TrackedMapState): TrackedMapState {
+    for (const deleteActionPayload of this._data) {
+      const plant = state.layers.plants.loadedObjects.find(
+        (obj) => obj.id === deleteActionPayload.id,
+      );
+      if (plant?.addDate) {
+        decreaseAddedPlantsForDate(plant.addDate);
+      }
+    }
+
     return {
       ...state,
       layers: {
@@ -293,6 +311,15 @@ export class UpdateAddDatePlantAction
 
     const timelineDate = useMapStore.getState().untrackedState.timelineDate;
 
+    for (const addDateActionPayload of this._data) {
+      const plant = state.layers.plants.loadedObjects.find(
+        (obj) => obj.id === addDateActionPayload.id,
+      );
+      if (plant?.addDate && addDateActionPayload.addDate) {
+        timlineEventsUpdateAdedDate(plant.addDate, addDateActionPayload.addDate);
+      }
+    }
+
     return {
       ...state,
       layers: {
@@ -311,6 +338,71 @@ export class UpdateAddDatePlantAction
 
   execute(mapId: number): Promise<PlantingDto> {
     return updateAddDatePlanting(mapId, this.actionId, this._data);
+  }
+}
+
+export class UpdatePlantingNotesAction
+  implements
+    Action<
+      Awaited<ReturnType<typeof updatePlantingNotes>>,
+      Awaited<ReturnType<typeof updatePlantingNotes>>
+    >
+{
+  private readonly _ids: string[];
+
+  constructor(private readonly _data: UpdatePlantingNoteActionPayload[], public actionId = v4()) {
+    this._ids = this._data.map(({ id }) => id);
+  }
+
+  get entityIds() {
+    return this._ids;
+  }
+
+  reverse(state: TrackedMapState) {
+    const plants = filterByIds(state.layers.plants.loadedObjects, this._ids);
+
+    if (!plants.length) {
+      return null;
+    }
+
+    return new UpdatePlantingNotesAction(
+      plants.map((p) => ({
+        id: p.id,
+        notes: p.plantingNotes || '',
+      })),
+      this.actionId,
+    );
+  }
+
+  apply(state: TrackedMapState): TrackedMapState {
+    const updatePlants = (plants: Array<PlantingDto>) => {
+      return plants.map((p) => {
+        if (this._ids.includes(p.id)) {
+          return {
+            ...p,
+            plantingNotes: this._data.find((d) => d.id === p.id)?.notes ?? p.plantingNotes,
+          };
+        }
+
+        return p;
+      });
+    };
+
+    return {
+      ...state,
+      layers: {
+        ...state.layers,
+        plants: {
+          ...state.layers.plants,
+          objects: updatePlants(state.layers.plants.objects),
+          loadedObjects: updatePlants(state.layers.plants.loadedObjects),
+        },
+      },
+    };
+  }
+
+  execute(mapId: number): Promise<PlantingDto> {
+    return updatePlantingNotes(mapId, this.actionId, this._data);
   }
 }
 
@@ -362,6 +454,13 @@ export class UpdateRemoveDatePlantAction
     };
 
     const timelineDate = useMapStore.getState().untrackedState.timelineDate;
+
+    for (const removeDateActionPayload of this._data) {
+      const plant = state.layers.plants.loadedObjects.find(
+        (obj) => obj.id === removeDateActionPayload.id,
+      );
+      timlineEventsUpdateRemoveDate(plant?.removeDate, removeDateActionPayload.removeDate);
+    }
 
     return {
       ...state,